fix: judge stale peers against our own validated tip, in 64 bits
Review caught two faults in the eviction rule. Judging peers against the
best height any peer claimed let one peer claiming Int32.max evict every
honest peer and their replacements until it held the pool alone; and the
Int32 subtraction trapped on a hostile Int32.min. The reference is now
the height of the header chain the pool itself synced — proof-of-work
checked, so no peer can inflate it — and the arithmetic is widened.

That reference exposed a second gap: header sync always asked the first
seated peer and stopped after one success, so a stale peer seated first
froze the tip and could never be judged. Header sync now also asks any
other peer claiming a tip well above ours, one round trip settled by
proof of work, so an honest peer's claim advances the tip and the stale
one is unseated on the next check.

Tests: the stale peer is unseated once headers have synced; a peer
claiming Int32.max cannot get honest peers evicted and one claiming
Int32.min does not trap; the synthetic chain builder is inclusive, so a
120-block chain tips at 120.

+++ Tests/BitcoinP2PTests/FallbackPeerGeneratorTests.swift
@@ -117,12 +117,12 @@ struct FallbackPeerGeneratorTests {
         // dead fork, and the pool would only unseat it at first contact
         // (`PeerPool.staleTipTolerance`). Shipping it wastes the user's first
         // dials on a peer known to be useless when the list was made.
-        let stale = verified.filter { tip - $0.startHeight > PeerPool.staleTipTolerance }
+        let stale = verified.filter { Int64(tip) - Int64($0.startHeight) > PeerPool.staleTipTolerance }
         for peer in stale {
             print("generator: dropping \(peer.endpoint) at height \(peer.startHeight), "
                   + "\(tip - peer.startHeight) behind the median \(peer.userAgent)")
         }
-        verified.removeAll { tip - $0.startHeight > PeerPool.staleTipTolerance }
+        verified.removeAll { Int64(tip) - Int64($0.startHeight) > PeerPool.staleTipTolerance }
         try #require(verified.count >= Self.floor,
                      "only \(verified.count) peers near the tip — refusing to ship a thin list")
         let date = ISO8601DateFormatter().string(from: Date())
